project_quotes/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djongo==1.3.6
dnspython==2.6.1
mongo==0.2.0
mongoengine==0.28.2
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pymongo==4.7.3
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
